refactor: clean up unused code after SPA migration
* React components are now universally in a `components` directory,
  since we have much fewer of them and this makes them easier to pick
  out among non-React code.

* Some code shared between frontend files was untangled and placed in a
  `daysOfWeek` helper file.

* Unused frontend packages were detected using the `depcheck` tool.
