+++ assets/css/app.css
@@ -6,13 +6,10 @@
 
 @import "./base.css";
 @import "./checkbox.css";
-@import "./icon.css";
-@import "./header.css";
-@import "./forms.css";
-@import "./disruption_times.css";
+@import "./datepicker.css";
 @import "./disruption_details.css";
 @import "./disruption_form.css";
-@import "./disruption_table.css";
 @import "./disruption_index.css";
-@import "./disruption_view_toggle.css";
-@import "./datepicker.css";
+@import "./disruption_table.css";
+@import "./header.css";
+@import "./icon.css";
