+++ assets/css/base.css
@@ -44,3 +44,23 @@ a:hover {
 h4 {
   font-weight: 700;
 }
+
+fieldset {
+  margin-left: 1.5rem;
+  margin-top: 1rem;
+}
+
+fieldset fieldset {
+  margin-left: 0;
+  margin-top: 0.5rem;
+}
+
+legend {
+  font-weight: bold;
+  margin-left: -1.5rem;
+}
+
+fieldset fieldset legend {
+  font-size: 100%;
+  margin-left: 0;
+}
