+++ assets/css/checkbox.css
@@ -2,6 +2,7 @@
   display: flex;
   align-items: center;
 }
+
 .m-checkbox input {
   position: absolute;
   width: 1px;
