+++ assets/css/disruption_details.css
@@ -23,41 +23,6 @@
   list-style: none;
 }
 
-.m-disruption-details__view-toggle-group {
-  min-width: 350px;
-  max-width: 50%;
-}
-
-.m-disruption-details__view-toggle {
-  border: 1px solid var(--primary);
-  border-bottom: none;
-  color: var(--black);
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 1rem;
-}
-
-.m-disruption-details__view-toggle:hover {
-  background-color: var(--light-pink);
-  color: var(--black);
-}
-
-.m-disruption-details__view-toggle.active {
-  background-color: var(--light-pink);
-}
-
-.m-disruption-details__view-toggle:first-child {
-  border-top-left-radius: 0.25rem;
-  border-top-right-radius: 0.25rem;
-}
-
-.m-disruption-details__view-toggle:last-child {
-  border-bottom: 1px solid var(--primary);
-  border-bottom-left-radius: 0.25rem;
-  border-bottom-right-radius: 0.25rem;
-}
-
 .m-disruption-details__deletion-indicator {
   background-color: var(--light-grey);
   border: 1px solid var(--blue-grey);
