+++ assets/css/disruption_index.css
@@ -1,6 +1,7 @@
 .m-disruption-index__route_filter {
   opacity: 0.3;
 }
+
 .m-disruption-index__route_filter.active {
   opacity: 1;
 }
