--- assets/css/disruption_times.css
@@ -1,11 +1,0 @@
-.m-disruption-times__dow_label {
-  width: 7rem;
-}
-
-.m-disruption-times__time_of_day_start {
-  margin-right: 1rem;
-}
-
-.m-disruption-times__time_of_day_end {
-  margin-left: 1rem;
-}
