--- assets/css/disruption_view_toggle.css
@@ -1,21 +1,0 @@
-.m-disruption-view-toggle {
-  display: flex;
-  flex-direction: column;
-}
-
-.m-disruption-view-toggle_button {
-  background-color: #f8f9fa;
-  border-color: #f8f9fa;
-}
-
-.m-disruption-view-toggle_button:first-child {
-  border-radius: 0.25rem 0.25rem 0 0;
-}
-
-.m-disruption-view-toggle_button:last-child {
-  border-radius: 0 0 0.25rem 0.25rem;
-}
-
-.m-disruption-view-toggle_button.active {
-  background-color: #cfe3f4;
-}
