--- assets/css/forms.css
@@ -1,38 +1,0 @@
-fieldset {
-  margin-left: 1.5rem;
-  margin-top: 1rem;
-}
-
-fieldset fieldset {
-  margin-left: 0;
-  margin-top: 0.5rem;
-}
-
-legend {
-  font-weight: bold;
-  margin-left: -1.5rem;
-}
-
-fieldset fieldset legend {
-  font-size: 100%;
-  margin-left: 0;
-}
-
-.m-forms__day-of-week-bubble input {
-  display: none;
-}
-
-.m-forms__day-of-week-bubble label {
-  border: 1px solid #dd3159;
-  border-radius: 1.25rem;
-  color: #dd3159;
-  height: 2rem;
-  padding-top: 0.2rem;
-  text-align: center;
-  width: 4rem;
-}
-
-.m-forms__day-of-week-bubble input:checked + label {
-  background-color: #dd3159;
-  color: white;
-}
