+++ assets/src/components/TimePicker.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react"
-import Checkbox from "./checkbox"
+import Checkbox from "./Checkbox"
 
 const hours = ["12", "1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11"]
 const minutes = ["00", "15", "30", "45"]
