+++ assets/src/daysOfWeek.ts
@@ -1,0 +1,20 @@
+type DayName =
+  | "monday"
+  | "tuesday"
+  | "wednesday"
+  | "thursday"
+  | "friday"
+  | "saturday"
+  | "sunday"
+
+const dayNumbers: { [name in DayName]: number } = {
+  monday: 0,
+  tuesday: 1,
+  wednesday: 2,
+  thursday: 3,
+  friday: 4,
+  saturday: 5,
+  sunday: 6,
+}
+
+export { DayName, dayNumbers }
