--- assets/src/disruptions/disruptionIndex.tsx
@@ -1,67 +1,0 @@
-import Icon from "../icons"
-
-const getRouteIcon = (route?: string): Icon => {
-  switch (route) {
-    case "Red": {
-      return "red-line-small"
-    }
-    case "Blue": {
-      return "blue-line-small"
-    }
-    case "Mattapan": {
-      return "mattapan-line-small"
-    }
-    case "Orange": {
-      return "orange-line-small"
-    }
-    case "Green-B": {
-      return "green-line-b-small"
-    }
-    case "Green-C": {
-      return "green-line-c-small"
-    }
-    case "Green-D": {
-      return "green-line-d-small"
-    }
-    case "Green-E": {
-      return "green-line-e-small"
-    }
-    default: {
-      return "mode-commuter-rail-small"
-    }
-  }
-}
-
-const getRouteColor = (route?: string): string => {
-  switch (route) {
-    case "Red": {
-      return "#da291c"
-    }
-    case "Blue": {
-      return "#003da5"
-    }
-    case "Mattapan": {
-      return "#da291c"
-    }
-    case "Orange": {
-      return "#ed8b00"
-    }
-    case "Green-B": {
-      return "#00843d"
-    }
-    case "Green-C": {
-      return "#00843d"
-    }
-    case "Green-D": {
-      return "#00843d"
-    }
-    case "Green-E": {
-      return "#00843d"
-    }
-    default: {
-      return "#80276c"
-    }
-  }
-}
-
-export { getRouteIcon, getRouteColor }
