--- assets/src/navigation.ts
@@ -1,5 +1,0 @@
-const redirectTo = (newLocation: string) => {
-  window.location.href = newLocation
-}
-
-export { redirectTo }
