--- lib/arrow_web/templates/page/index.html.eex
@@ -1,2 +1,0 @@
-<div id="app"></div>
-<script type="text/javascript" src="<%= Routes.static_path(@conn, "/js/app.js") %>"></script>
